refactor: move globals to registry. Also some registry components related
