+++ web-registry/src/components/organisms/BridgeModal/BridgeModal.tsx
@@ -1,12 +1,13 @@
-import {
-  BridgeForm,
-  BridgeProps,
-} from 'web-components/lib/components/form/BridgeForm';
 import { ToucanIcon } from 'web-components/lib/components/icons/ToucanIcon';
 import { RegenModalProps } from 'web-components/lib/components/modal';
 import { FormModalTemplate } from 'web-components/lib/components/modal/FormModalTemplate';
 
 import { BatchInfoWithBalance } from 'types/ledger/ecocredit';
+
+import {
+  BridgeForm,
+  BridgeProps,
+} from 'components/organisms/BridgeForm/BridgeForm';
 
 import {
   BRIDGE_MODAL_SUBTITLE,
