+++ web-registry/src/config/globals.ts
@@ -7,3 +7,6 @@ export const URL_REGISTRY_TERMS_SERVICE =
   'https://regennetwork.notion.site/Platform-Terms-of-Service-b77faf978cd04e2e8d3c58f76841bad1';
 export const URL_REGISTRY_MARKETPLACE_LEGAL =
   'https://regennetwork.notion.site/Regen-Marketplace-Contracts-7bc2b0075ee64c1a957463bb5eddcead';
+
+// Env vars
+export const MAPBOX_TOKEN = process.env.REACT_APP_MAPBOX_TOKEN || '';
